LPD-92286 refactor: extract report feedback logic into a useReportFeedback hook
