+++ modules/apps/ai-hub-cell/ai-hub-cell-js-components-web/src/main/resources/META-INF/resources/js/ReportFeedback/ReportFeedbackModal.tsx
@@ -8,14 +8,11 @@ import ClayButton from '@clayui/button';
 import ClayForm, {ClayInput, ClaySelectWithOption} from '@clayui/form';
 import ClayModal, {useModal} from '@clayui/modal';
 import {RequiredMask} from 'frontend-js-components-web';
-import React, {useState} from 'react';
+import React from 'react';
 
-import {
-	ReportFeedbackPayload,
-	ReportFeedbackReason,
-	ReportFeedbackSurface,
-	postAIIssueReport,
-} from './api';
+import {ReportFeedbackReason, ReportFeedbackSurface} from './api';
+import {REASON_OPTIONS} from './constants';
+import useReportFeedback from './useReportFeedback';
 
 interface ReportFeedbackModalProps {
 	agentId: string;
@@ -26,30 +23,6 @@ interface ReportFeedbackModalProps {
 	traceId: string;
 }
 
-const REASON_OPTIONS: {label: string; value: ReportFeedbackReason | ''}[] = [
-	{label: Liferay.Language.get('select-reason'), value: ''},
-	{
-		label: Liferay.Language.get('incorrect-or-inaccurate-response'),
-		value: 'INCORRECT_OR_INACCURATE_RESPONSE',
-	},
-	{
-		label: Liferay.Language.get('inappropriate-or-harmful-content'),
-		value: 'INAPPROPRIATE_OR_HARMFUL_CONTENT',
-	},
-	{
-		label: Liferay.Language.get('exposure-of-personal-sensitive-data-pii'),
-		value: 'PII_EXPOSURE',
-	},
-	{
-		label: Liferay.Language.get('agent-error-or-malfunction'),
-		value: 'AGENT_ERROR_OR_MALFUNCTION',
-	},
-	{
-		label: Liferay.Language.get('other'),
-		value: 'OTHER',
-	},
-];
-
 const ReportFeedbackModal: React.FC<ReportFeedbackModalProps> = ({
 	agentId,
 	agentVersion,
@@ -58,37 +31,23 @@ const ReportFeedbackModal: React.FC<ReportFeedbackModalProps> = ({
 	surface,
 	traceId,
 }) => {
-	const [reason, setReason] = useState<ReportFeedbackReason | ''>('');
-	const [comment, setComment] = useState<string>('');
-	const [submitting, setSubmitting] = useState<boolean>(false);
-	const [error, setError] = useState<string>('');
-
 	const {observer, onClose: closeModal} = useModal({onClose});
 
-	const canSubmit = Boolean(reason) && !submitting;
+	const {
+		canSubmit,
+		comment,
+		error,
+		reason,
+		setComment,
+		setReason,
+		submit,
+		submitting,
+	} = useReportFeedback({agentId, agentVersion, surface, traceId});
 
 	async function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
 		event.preventDefault();
 
-		if (!reason) {
-			return;
-		}
-
-		const payload: ReportFeedbackPayload = {
-			agentId,
-			agentVersion,
-			reason,
-			surface,
-			traceId,
-			...(comment.trim() ? {comment: comment.trim()} : {}),
-		};
-
-		setError('');
-		setSubmitting(true);
-
-		try {
-			await postAIIssueReport(payload);
-
+		if (await submit()) {
 			Liferay.Util.openToast({
 				message: Liferay.Language.get(
 					'feedback-sent-thank-you-for-helping-us-improve'
@@ -98,15 +57,6 @@ const ReportFeedbackModal: React.FC<ReportFeedbackModalProps> = ({
 
 			onSubmitted?.();
 			closeModal();
-		}
-		catch (caught) {
-			setError(
-				(caught as Error).message ||
-					Liferay.Language.get(
-						'unable-to-send-feedback-please-try-again'
-					)
-			);
-			setSubmitting(false);
 		}
 	}
